# AsyncJobsTemplate.WebApi/OpenApi/async-jobs-template.yaml
openapi: 3.0.1
info:
  title: AsyncJobTemplate
  version: '1.0'
paths:
...
  '/jobs/{jobCategoryName}':
    post:
      tags:
        - Jobs
      summary: Trigger a job with a JSON payload
      parameters:
        - name: jobCategoryName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema: { }
          text/json:
            schema: { }
          application/*+json:
            schema: { }
      responses:
        '200':
          description: Correct response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriggerJobResponse'
        '400':
          description: Incorrect request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriggerJobResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriggerJobResponse'
  '/jobs/{jobCategoryName}/file':
    post:
      tags:
        - Jobs
      summary: Trigger a job with a file
      parameters:
        - name: jobCategoryName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
            encoding:
              file:
                style: form
      responses:
        '200':
          description: Correct response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriggerJobResponse'
        '400':
          description: Incorrect request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriggerJobResponse'
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriggerJobResponse'
  '/jobs/{jobId}':
    get:
      tags:
        - Jobs
      summary: Get a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Correct response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetJobResult'
        '404':
          description: Job doesn't exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetJobResult'
        '500':
          description: Internal Server Error
  /jobs:
    get:
      tags:
        - Jobs
      summary: Get jobs
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            format: int32
        - name: pageSize
          in: query
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Correct response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetJobsResult'
  '/jobs/{jobId}/file':
    get:
      tags:
        - Jobs
      summary: Download a job file
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '404':
          description: Job file doesn't exist
components:
  schemas:
    Assembly:
      type: object
      properties:
        definedTypes:
          type: array
          items:
            $ref: '#/components/schemas/TypeInfo'
          nullable: true
          readOnly: true
        exportedTypes:
          type: array
          items:
            $ref: '#/components/schemas/Type'
          nullable: true
          readOnly: true
        codeBase:
          type: string
          nullable: true
          readOnly: true
          deprecated: true
        entryPoint:
          $ref: '#/components/schemas/MethodInfo'
        fullName:
          type: string
          nullable: true
          readOnly: true
        imageRuntimeVersion:
          type: string
          nullable: true
          readOnly: true
        isDynamic:
          type: boolean
          readOnly: true
        location:
          type: string
          nullable: true
          readOnly: true
        reflectionOnly:
          type: boolean
          readOnly: true
        isCollectible:
          type: boolean
          readOnly: true
        isFullyTrusted:
          type: boolean
          readOnly: true
        customAttributes:
          type: array
          items:
            $ref: '#/components/schemas/CustomAttributeData'
          nullable: true
          readOnly: true
        escapedCodeBase:
          type: string
          nullable: true
          readOnly: true
          deprecated: true
        manifestModule:
          $ref: '#/components/schemas/Module'
        modules:
          type: array
          items:
            $ref: '#/components/schemas/Module'
          nullable: true
          readOnly: true
        globalAssemblyCache:
          type: boolean
          readOnly: true
          deprecated: true
        hostContext:
          type: integer
          format: int64
          readOnly: true
        securityRuleSet:
          $ref: '#/components/schemas/SecurityRuleSet'
      additionalProperties: false
    CallingConventions:
      enum:
        - 1
        - 2
        - 3
        - 32
        - 64
      type: integer
      format: int32
    ConstructorInfo:
      type: object
      properties:
        name:
          type: string
          nullable: true
          readOnly: true
        declaringType:
          $ref: '#/components/schemas/Type'
        reflectedType:
          $ref: '#/components/schemas/Type'
        module:
          $ref: '#/components/schemas/Module'
        customAttributes:
          type: array
          items:
            $ref: '#/components/schemas/CustomAttributeData'
          nullable: true
          readOnly: true
        isCollectible:
          type: boolean
          readOnly: true
        metadataToken:
          type: integer
          format: int32
          readOnly: true
        attributes:
          $ref: '#/components/schemas/MethodAttributes'
        methodImplementationFlags:
          $ref: '#/components/schemas/MethodImplAttributes'
        callingConvention:
          $ref: '#/components/schemas/CallingConventions'
        isAbstract:
          type: boolean
          readOnly: true
        isConstructor:
          type: boolean
          readOnly: true
        isFinal:
          type: boolean
          readOnly: true
        isHideBySig:
          type: boolean
          readOnly: true
        isSpecialName:
          type: boolean
          readOnly: true
        isStatic:
          type: boolean
          readOnly: true
        isVirtual:
          type: boolean
          readOnly: true
        isAssembly:
          type: boolean
          readOnly: true
        isFamily:
          type: boolean
          readOnly: true
        isFamilyAndAssembly:
          type: boolean
          readOnly: true
        isFamilyOrAssembly:
          type: boolean
          readOnly: true
        isPrivate:
          type: boolean
          readOnly: true
        isPublic:
          type: boolean
          readOnly: true
        isConstructedGenericMethod:
          type: boolean
          readOnly: true
        isGenericMethod:
          type: boolean
          readOnly: true
        isGenericMethodDefinition:
          type: boolean
          readOnly: true
        containsGenericParameters:
          type: boolean
          readOnly: true
        methodHandle:
          $ref: '#/components/schemas/RuntimeMethodHandle'
        isSecurityCritical:
          type: boolean
          readOnly: true
        isSecuritySafeCritical:
          type: boolean
          readOnly: true
        isSecurityTransparent:
          type: boolean
          readOnly: true
        memberType:
          $ref: '#/components/schemas/MemberTypes'
      additionalProperties: false
    CustomAttributeData:
      type: object
      properties:
        attributeType:
          $ref: '#/components/schemas/Type'
        constructor:
          $ref: '#/components/schemas/ConstructorInfo'
        constructorArguments:
          type: array
          items:
            $ref: '#/components/schemas/CustomAttributeTypedArgument'
          nullable: true
          readOnly: true
        namedArguments:
          type: array
          items:
            $ref: '#/components/schemas/CustomAttributeNamedArgument'
          nullable: true
          readOnly: true
      additionalProperties: false
    CustomAttributeNamedArgument:
      type: object
      properties:
        memberInfo:
          $ref: '#/components/schemas/MemberInfo'
        typedValue:
          $ref: '#/components/schemas/CustomAttributeTypedArgument'
        memberName:
          type: string
          nullable: true
          readOnly: true
        isField:
          type: boolean
          readOnly: true
      additionalProperties: false
    CustomAttributeTypedArgument:
      type: object
      properties:
        argumentType:
          $ref: '#/components/schemas/Type'
        value:
          nullable: true
      additionalProperties: false
    ErrorType:
      enum:
        - 0
        - 1
      type: integer
      format: int32
...
    Exception:
      type: object
      properties:
        targetSite:
          $ref: '#/components/schemas/MethodBase'
        message:
          type: string
          nullable: true
          readOnly: true
        data:
          type: object
          additionalProperties: { }
          nullable: true
          readOnly: true
        innerException:
          $ref: '#/components/schemas/Exception'
        helpLink:
          type: string
          nullable: true
        source:
          type: string
          nullable: true
        hResult:
          type: integer
          format: int32
        stackTrace:
          type: string
          nullable: true
          readOnly: true
      additionalProperties: false
...
    GenericParameterAttributes:
      enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 8
        - 16
        - 28
        - 32
      type: integer
      format: int32
...
    GetJobResult:
      type: object
      properties:
        status:
          type: string
          nullable: true
        outputData:
          nullable: true
        outputFileReference:
          type: string
          nullable: true
      additionalProperties: false
    GetJobsResult:
      type: object
      properties:
        jobs:
          $ref: '#/components/schemas/JobPaginatedList'
      additionalProperties: false
    ICustomAttributeProvider:
      type: object
      additionalProperties: false
    IntPtr:
      type: object
      additionalProperties: false
    Job:
      type: object
      properties:
        jobId:
          type: string
          format: uuid
        jobCategoryName:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/JobStatus'
        inputData:
          nullable: true
        inputFileReference:
          type: string
          nullable: true
        outputData:
          nullable: true
        outputFileReference:
          type: string
          nullable: true
        errors:
          type: array
          items:
            $ref: '#/components/schemas/JobError'
          nullable: true
        createdAtUtc:
          type: string
          format: date-time
        lastUpdatedAtUtc:
          type: string
          format: date-time
          nullable: true
      additionalProperties: false
    JobError:
      type: object
      properties:
        message:
          type: string
          nullable: true
        exception:
          $ref: '#/components/schemas/Exception'
        errorCode:
          type: string
          nullable: true
        type:
          $ref: '#/components/schemas/ErrorType'
      additionalProperties: false
    JobPaginatedList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Job'
          nullable: true
        count:
          type: integer
          format: int64
      additionalProperties: false
    JobStatus:
      enum:
        - 0
        - 1
        - 2
        - 3
        - 4
      type: integer
      format: int32
    LayoutKind:
      enum:
        - 0
        - 2
        - 3
      type: integer
      format: int32
    MemberInfo:
      type: object
      properties:
        memberType:
          $ref: '#/components/schemas/MemberTypes'
        name:
          type: string
          nullable: true
          readOnly: true
        declaringType:
          $ref: '#/components/schemas/Type'
        reflectedType:
          $ref: '#/components/schemas/Type'
        module:
          $ref: '#/components/schemas/Module'
        customAttributes:
          type: array
          items:
            $ref: '#/components/schemas/CustomAttributeData'
          nullable: true
          readOnly: true
        isCollectible:
          type: boolean
          readOnly: true
        metadataToken:
          type: integer
          format: int32
          readOnly: true
      additionalProperties: false
    MemberTypes:
      enum:
        - 1
        - 2
        - 4
        - 8
        - 16
        - 32
        - 64
        - 128
        - 191
      type: integer
      format: int32
    MethodAttributes:
      enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        - 7
        - 8
        - 16
        - 32
        - 64
        - 128
        - 256
        - 512
        - 1024
        - 2048
        - 4096
        - 8192
        - 16384
        - 32768
        - 53248
      type: integer
      format: int32
    MethodBase:
      type: object
      properties:
        memberType:
          $ref: '#/components/schemas/MemberTypes'
        name:
          type: string
          nullable: true
          readOnly: true
        declaringType:
          $ref: '#/components/schemas/Type'
        reflectedType:
          $ref: '#/components/schemas/Type'
        module:
          $ref: '#/components/schemas/Module'
        customAttributes:
          type: array
          items:
            $ref: '#/components/schemas/CustomAttributeData'
          nullable: true
          readOnly: true
        isCollectible:
          type: boolean
          readOnly: true
        metadataToken:
          type: integer
          format: int32
          readOnly: true
        attributes:
          $ref: '#/components/schemas/MethodAttributes'
        methodImplementationFlags:
          $ref: '#/components/schemas/MethodImplAttributes'
        callingConvention:
          $ref: '#/components/schemas/CallingConventions'
        isAbstract:
          type: boolean
          readOnly: true
        isConstructor:
          type: boolean
          readOnly: true
        isFinal:
          type: boolean
          readOnly: true
        isHideBySig:
          type: boolean
          readOnly: true
        isSpecialName:
          type: boolean
          readOnly: true
        isStatic:
          type: boolean
          readOnly: true
        isVirtual:
          type: boolean
          readOnly: true
        isAssembly:
          type: boolean
          readOnly: true
        isFamily:
          type: boolean
          readOnly: true
        isFamilyAndAssembly:
          type: boolean
          readOnly: true
        isFamilyOrAssembly:
          type: boolean
          readOnly: true
        isPrivate:
          type: boolean
          readOnly: true
        isPublic:
          type: boolean
          readOnly: true
        isConstructedGenericMethod:
          type: boolean
          readOnly: true
        isGenericMethod:
          type: boolean
          readOnly: true
        isGenericMethodDefinition:
          type: boolean
          readOnly: true
        containsGenericParameters:
          type: boolean
          readOnly: true
        methodHandle:
          $ref: '#/components/schemas/RuntimeMethodHandle'
        isSecurityCritical:
          type: boolean
          readOnly: true
        isSecuritySafeCritical:
          type: boolean
          readOnly: true
        isSecurityTransparent:
          type: boolean
          readOnly: true
      additionalProperties: false
    MethodImplAttributes:
      enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 8
        - 16
        - 32
        - 64
        - 128
        - 256
        - 512
        - 4096
        - 65535
      type: integer
      format: int32
    MethodInfo:
      type: object
      properties:
        name:
          type: string
          nullable: true
          readOnly: true
        declaringType:
          $ref: '#/components/schemas/Type'
        reflectedType:
          $ref: '#/components/schemas/Type'
        module:
          $ref: '#/components/schemas/Module'
        customAttributes:
          type: array
          items:
            $ref: '#/components/schemas/CustomAttributeData'
          nullable: true
          readOnly: true
        isCollectible:
          type: boolean
          readOnly: true
        metadataToken:
          type: integer
          format: int32
          readOnly: true
        attributes:
          $ref: '#/components/schemas/MethodAttributes'
        methodImplementationFlags:
          $ref: '#/components/schemas/MethodImplAttributes'
        callingConvention:
          $ref: '#/components/schemas/CallingConventions'
        isAbstract:
          type: boolean
          readOnly: true
        isConstructor:
          type: boolean
          readOnly: true
        isFinal:
          type: boolean
          readOnly: true
        isHideBySig:
          type: boolean
          readOnly: true
        isSpecialName:
          type: boolean
          readOnly: true
        isStatic:
          type: boolean
          readOnly: true
        isVirtual:
          type: boolean
          readOnly: true
        isAssembly:
          type: boolean
          readOnly: true
        isFamily:
          type: boolean
          readOnly: true
        isFamilyAndAssembly:
          type: boolean
          readOnly: true
        isFamilyOrAssembly:
          type: boolean
          readOnly: true
        isPrivate:
          type: boolean
          readOnly: true
        isPublic:
          type: boolean
          readOnly: true
        isConstructedGenericMethod:
          type: boolean
          readOnly: true
        isGenericMethod:
          type: boolean
          readOnly: true
        isGenericMethodDefinition:
          type: boolean
          readOnly: true
        containsGenericParameters:
          type: boolean
          readOnly: true
        methodHandle:
          $ref: '#/components/schemas/RuntimeMethodHandle'
        isSecurityCritical:
          type: boolean
          readOnly: true
        isSecuritySafeCritical:
          type: boolean
          readOnly: true
        isSecurityTransparent:
          type: boolean
          readOnly: true
        memberType:
          $ref: '#/components/schemas/MemberTypes'
        returnParameter:
          $ref: '#/components/schemas/ParameterInfo'
        returnType:
          $ref: '#/components/schemas/Type'
        returnTypeCustomAttributes:
          $ref: '#/components/schemas/ICustomAttributeProvider'
      additionalProperties: false
    Module:
      type: object
      properties:
        assembly:
          $ref: '#/components/schemas/Assembly'
        fullyQualifiedName:
          type: string
          nullable: true
          readOnly: true
        name:
          type: string
          nullable: true
          readOnly: true
        mdStreamVersion:
          type: integer
          format: int32
          readOnly: true
        moduleVersionId:
          type: string
          format: uuid
          readOnly: true
        scopeName:
          type: string
          nullable: true
          readOnly: true
        moduleHandle:
          $ref: '#/components/schemas/ModuleHandle'
        customAttributes:
          type: array
          items:
            $ref: '#/components/schemas/CustomAttributeData'
          nullable: true
          readOnly: true
        metadataToken:
          type: integer
          format: int32
          readOnly: true
      additionalProperties: false
    ModuleHandle:
      type: object
      properties:
        mdStreamVersion:
          type: integer
          format: int32
          readOnly: true
      additionalProperties: false
    ParameterAttributes:
      enum:
        - 0
        - 1
        - 2
        - 4
        - 8
        - 16
        - 4096
        - 8192
        - 16384
        - 32768
        - 61440
      type: integer
      format: int32
    ParameterInfo:
      type: object
      properties:
        attributes:
          $ref: '#/components/schemas/ParameterAttributes'
        member:
          $ref: '#/components/schemas/MemberInfo'
        name:
          type: string
          nullable: true
          readOnly: true
        parameterType:
          $ref: '#/components/schemas/Type'
        position:
          type: integer
          format: int32
          readOnly: true
        isIn:
          type: boolean
          readOnly: true
        isLcid:
          type: boolean
          readOnly: true
        isOptional:
          type: boolean
          readOnly: true
        isOut:
          type: boolean
          readOnly: true
        isRetval:
          type: boolean
          readOnly: true
        defaultValue:
          nullable: true
          readOnly: true
        rawDefaultValue:
          nullable: true
          readOnly: true
        hasDefaultValue:
          type: boolean
          readOnly: true
        customAttributes:
          type: array
          items:
            $ref: '#/components/schemas/CustomAttributeData'
          nullable: true
          readOnly: true
        metadataToken:
          type: integer
          format: int32
          readOnly: true
      additionalProperties: false
...
    RuntimeMethodHandle:
      type: object
      properties:
        value:
          $ref: '#/components/schemas/IntPtr'
      additionalProperties: false
    RuntimeTypeHandle:
      type: object
      properties:
        value:
          $ref: '#/components/schemas/IntPtr'
      additionalProperties: false
    SecurityRuleSet:
      enum:
        - 0
        - 1
        - 2
      type: integer
      format: int32
    StructLayoutAttribute:
      type: object
      properties:
        typeId:
          nullable: true
          readOnly: true
        value:
          $ref: '#/components/schemas/LayoutKind'
      additionalProperties: false
    TriggerJobResponse:
      type: object
      properties:
        result:
          type: boolean
        jobId:
          type: string
          format: uuid
          nullable: true
      additionalProperties: false
    Type:
      type: object
      properties:
        name:
          type: string
          nullable: true
          readOnly: true
        customAttributes:
          type: array
          items:
            $ref: '#/components/schemas/CustomAttributeData'
          nullable: true
          readOnly: true
        isCollectible:
          type: boolean
          readOnly: true
        metadataToken:
          type: integer
          format: int32
          readOnly: true
        memberType:
          $ref: '#/components/schemas/MemberTypes'
        namespace:
          type: string
          nullable: true
          readOnly: true
        assemblyQualifiedName:
          type: string
          nullable: true
          readOnly: true
        fullName:
          type: string
          nullable: true
          readOnly: true
        assembly:
          $ref: '#/components/schemas/Assembly'
        module:
          $ref: '#/components/schemas/Module'
        isInterface:
          type: boolean
          readOnly: true
        isNested:
          type: boolean
          readOnly: true
        declaringType:
          $ref: '#/components/schemas/Type'
        declaringMethod:
          $ref: '#/components/schemas/MethodBase'
        reflectedType:
          $ref: '#/components/schemas/Type'
        underlyingSystemType:
          $ref: '#/components/schemas/Type'
        isTypeDefinition:
          type: boolean
          readOnly: true
        isArray:
          type: boolean
          readOnly: true
        isByRef:
          type: boolean
          readOnly: true
        isPointer:
          type: boolean
          readOnly: true
        isConstructedGenericType:
          type: boolean
          readOnly: true
        isGenericParameter:
          type: boolean
          readOnly: true
        isGenericTypeParameter:
          type: boolean
          readOnly: true
        isGenericMethodParameter:
          type: boolean
          readOnly: true
        isGenericType:
          type: boolean
          readOnly: true
        isGenericTypeDefinition:
          type: boolean
          readOnly: true
        isSZArray:
          type: boolean
          readOnly: true
        isVariableBoundArray:
          type: boolean
          readOnly: true
        isByRefLike:
          type: boolean
          readOnly: true
        isFunctionPointer:
          type: boolean
          readOnly: true
        isUnmanagedFunctionPointer:
          type: boolean
          readOnly: true
        hasElementType:
          type: boolean
          readOnly: true
        genericTypeArguments:
          type: array
          items:
            $ref: '#/components/schemas/Type'
          nullable: true
          readOnly: true
        genericParameterPosition:
          type: integer
          format: int32
          readOnly: true
        genericParameterAttributes:
          $ref: '#/components/schemas/GenericParameterAttributes'
        attributes:
          $ref: '#/components/schemas/TypeAttributes'
        isAbstract:
          type: boolean
          readOnly: true
        isImport:
          type: boolean
          readOnly: true
        isSealed:
          type: boolean
          readOnly: true
        isSpecialName:
          type: boolean
          readOnly: true
        isClass:
          type: boolean
          readOnly: true
        isNestedAssembly:
          type: boolean
          readOnly: true
        isNestedFamANDAssem:
          type: boolean
          readOnly: true
        isNestedFamily:
          type: boolean
          readOnly: true
        isNestedFamORAssem:
          type: boolean
          readOnly: true
        isNestedPrivate:
          type: boolean
          readOnly: true
        isNestedPublic:
          type: boolean
          readOnly: true
        isNotPublic:
          type: boolean
          readOnly: true
        isPublic:
          type: boolean
          readOnly: true
        isAutoLayout:
          type: boolean
          readOnly: true
        isExplicitLayout:
          type: boolean
          readOnly: true
        isLayoutSequential:
          type: boolean
          readOnly: true
        isAnsiClass:
          type: boolean
          readOnly: true
        isAutoClass:
          type: boolean
          readOnly: true
        isUnicodeClass:
          type: boolean
          readOnly: true
        isCOMObject:
          type: boolean
          readOnly: true
        isContextful:
          type: boolean
          readOnly: true
        isEnum:
          type: boolean
          readOnly: true
        isMarshalByRef:
          type: boolean
          readOnly: true
        isPrimitive:
          type: boolean
          readOnly: true
        isValueType:
          type: boolean
          readOnly: true
        isSignatureType:
          type: boolean
          readOnly: true
        isSecurityCritical:
          type: boolean
          readOnly: true
        isSecuritySafeCritical:
          type: boolean
          readOnly: true
        isSecurityTransparent:
          type: boolean
          readOnly: true
        structLayoutAttribute:
          $ref: '#/components/schemas/StructLayoutAttribute'
        typeInitializer:
          $ref: '#/components/schemas/ConstructorInfo'
        typeHandle:
          $ref: '#/components/schemas/RuntimeTypeHandle'
        guid:
          type: string
          format: uuid
          readOnly: true
        baseType:
          $ref: '#/components/schemas/Type'
        isSerializable:
          type: boolean
          readOnly: true
          deprecated: true
        containsGenericParameters:
          type: boolean
          readOnly: true
        isVisible:
          type: boolean
          readOnly: true
      additionalProperties: false
    TypeAttributes:
      enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        - 7
        - 8
        - 16
        - 24
        - 32
        - 128
        - 256
        - 1024
        - 2048
        - 4096
        - 8192
        - 16384
        - 65536
        - 131072
        - 196608
        - 262144
        - 264192
        - 1048576
        - 12582912
      type: integer
      format: int32
...